+++ .github/workflows/ReleaseNotes.md
@@ -1,2 +1,3 @@
 * [Core] Fixed dead lock and race conditions in new _AsyncLock_ implementation (#1542).
 * [Client] Fixed connection freeze when using Xamarin etc.
+* [Server] Fixed issue in upgrading and downgrading of QoS levels for subscriptions and retained messages (#1560).